Adding SSL MITM Support.
We now can intercept SSL connections. We start an SSL client connection
with the server, as well as acting like an SSL server to the client.

We can now sniff and alter HTTPS traffic.
However, this (obviously) generates a client-side warning. The Java
application will have to add the appropriate certificates to Android.